First file move and rudimentary MQTT functions written
