+++ Ruby/Telegram/lib/Telegram.rb
@@ -1,0 +1,3 @@
+
+require_relative "Telegram/HTTPCore.rb"
+require_relative "Telegram/SingleUser.rb"
